+++ mlfz/functional/numpy.py
